refactor(glm): split HRF models into _hrf_models module
Move _gamma_difference_hrf, glover_hrf, spm_hrf, and _hrf_kernel from
_design.py into a dedicated _hrf_models module. Update test imports
accordingly. _design.py now focuses on design matrix assembly only.
